Implement version and help info printing
Write the option parsing code to handle version and help printing. Also
be sure to include the GPLv3 notice in each file. Also include a simple
Makefile to compile the program and clean the resulting binary.

Nothing really happens right now since arguments are ignored and only
the options are handled (to a simple degree).